requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
slack-bolt==1.18.0
slack-sdk==3.26.0
pydantic==2.4.2
python-multipart==0.0.6

